global: Change string position/length from unsigned int to size_t
Mainly to avoid truncating >4GB strings, which might potentially cause
some security holes. Normally there are other limits, which prevent such
excessive strings from being created in the first place.

I'm sure this didn't find everything. Maybe everything could be found with
compiler warnings. -Wconversion kind of does it, but it gives way too many
unnecessary warnings.

These were mainly found with:

grep " = strlen"
egrep "unsigned int.*(size|len)"

+++ src/auth/auth-cache.c
@@ -271,7 +271,7 @@ static bool auth_cache_node_is_user(struct auth_cache_node *node,
 				    const char *username)
 {
 	const char *data = node->data;
-	unsigned int username_len;
+	size_t username_len;
 
 	/* The cache nodes begin with "P"/"U", passdb/userdb ID, optional
 	   "+" master user, "\t" and then usually followed by the username.
